refactor: BattlePage join/resume 로직 훅 분리 (#246) - useBattleJoin 훅으로 세션 복구 및 재접속 로직 분리 - BattlePage에서 관련 effect 제거로 deps 단순화 - INVALID_ROLE 처리 콜백화로 역할 모달 처리 일원화
